+++ daskms/table_executor.py
@@ -21,6 +21,11 @@ STANDARD_EXECUTOR = "__standard_executor__"
 class ExecutorMetaClass(type):
     """ https://en.wikipedia.org/wiki/Multiton_pattern """
     def __call__(cls, key=STANDARD_EXECUTOR):
+        try:
+            return _executor_cache[key]
+        except KeyError:
+            pass
+
         with _executor_lock:
             try:
                 return _executor_cache[key]